requirements.txt
Django==1.11.5
Fabric3==1.13.1.post1
django-compressor==2.2
django-debug-toolbar==1.8
django-form-utils==1.0.3
django-libsass==0.7
django-npm==1.0.0
django-recaptcha==1.3.0
djangocms-admin-style==1.2.7
ipdb==0.10.3
ipython==6.0.0
model-mommy==1.3.2
psycopg2==2.7.1
pypugjs==4.2.2
pytz>=2017.2
django-loginas==0.3.2
